requirements.txt
streamlit==1.56.0
pandas==3.0.2
plotly==6.6.0
python-dotenv
google-api-python-client
